refactor(shared-session): shared_sessions 主体拆为四子模块（2061→954）
