+++ src-tauri/src/shared_sessions.rs
@@ -1,8 +1,5 @@
-use std::collections::{BTreeMap, HashMap};
-use std::io::Write;
-use std::path::{Path, PathBuf};
-use std::thread;
-use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+use std::collections::HashMap;
+use std::path::PathBuf;
 
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
@@ -10,21 +7,50 @@ use tauri::{AppHandle, State};
 use tokio::sync::Mutex;
 use uuid::Uuid;
 
-use crate::app_paths;
 use crate::codex;
 use crate::engine::{self, EngineType};
 use crate::shared::codex_core;
-use crate::shared_binding_visibility::collect_v2_shared_binding_ids_by_session;
-use crate::shared_event_log::{SessionTargetUpdate, SharedEventWriter};
 use crate::state::AppState;
 
-const SHARED_SESSIONS_DIRNAME: &str = "shared-sessions";
-const SHARED_STORE_LOCK_WAIT_TIMEOUT: Duration = Duration::from_secs(5);
-const SHARED_STORE_LOCK_RETRY_INTERVAL: Duration = Duration::from_millis(25);
-const SHARED_STORE_LOCK_STALE_TIMEOUT: Duration = Duration::from_secs(30);
-const MAX_DELTA_SYNC_TURNS: usize = 8;
-const MAX_DELTA_SYNC_CHARS: usize = 4_000;
-const SHARED_SESSION_SCHEMA_VERSION: u32 = 2;
+#[path = "shared_sessions/delta_sync.rs"]
+mod delta_sync;
+#[path = "shared_sessions/selection.rs"]
+mod selection;
+#[path = "shared_sessions/store.rs"]
+mod store;
+#[path = "shared_sessions/thread_id.rs"]
+mod thread_id;
+
+pub(crate) use delta_sync::{build_delta_sync_prefix, count_user_turns, extract_first_user_title};
+#[cfg(test)]
+pub(crate) use delta_sync::{inspect_shared_context_projection, MAX_DELTA_SYNC_CHARS};
+pub(crate) use selection::{
+    default_target_binding_availability, is_legacy_engine_only_selected_target,
+    normalize_provider_selection_source, normalize_shared_selected_target, select_meta_target,
+    upsert_v2_selected_target, validate_resolved_shared_selected_target,
+    write_shared_session_selection, SharedEngineBinding, SharedSelectedReasoning,
+    SharedTargetBindingMeta,
+};
+#[cfg(test)]
+pub(crate) use selection::{
+    apply_selected_target_selection, legacy_engine_only_selected_target,
+    resolve_shared_selection_update, sanitize_shared_session_meta, select_meta_engine_compat,
+};
+pub use selection::SharedSelectedTarget;
+pub(crate) use store::{
+    append_shared_session_log_entry, list_workspace_shared_sessions,
+    load_workspace_shared_ownership_seed, now_millis, read_latest_shared_session_snapshot,
+    read_shared_session_meta, shared_session_dir, shared_session_projection_source,
+    write_shared_session_meta, SharedSessionMeta, SharedSessionSnapshotEntry,
+    SharedSessionSummary, SHARED_SESSION_SCHEMA_VERSION,
+};
+#[cfg(test)]
+pub(crate) use store::load_seed_from_shared_sessions_dir;
+pub(crate) use thread_id::{
+    binding_uses_established_native_thread, canonical_shared_native_thread_id,
+    engine_binding_thread_id, is_pending_shared_binding_thread_id, parse_shared_session_id,
+    shared_target_binding_key, shared_thread_id, validate_shared_native_thread_id,
+};
 
 fn codex_turn_developer_instructions(settings: &crate::types::AppSettings) -> Option<String> {
     crate::backend::app_server_cli::codex_generated_developer_instructions_for_turn(settings)
@@ -64,218 +90,6 @@ pub(crate) fn ensure_supported_shared_session_engine(
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct SharedEngineBinding {
-    engine: EngineType,
-    native_thread_id: String,
-    created_at: u64,
-    last_used_at: u64,
-    last_synced_turn_seq: u64,
-}
-
-/// Target 级 Binding（Wave 4 / B.2）：Binding Key = Engine + ProviderProfile。
-/// `provider_profile_id = None` 表示 default/local Provider 语义（旧 V0 binding 的归位点）。
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct SharedTargetBindingMeta {
-    pub(crate) binding_key: String,
-    pub(crate) engine: EngineType,
-    pub(crate) provider_profile_id: Option<String>,
-    pub(crate) native_thread_id: String,
-    pub(crate) created_at: u64,
-    pub(crate) last_used_at: u64,
-    pub(crate) last_synced_turn_seq: u64,
-    /// ready / missing-provider / missing-runtime / degraded / recovery-required。
-    #[serde(default = "default_target_binding_availability")]
-    pub(crate) availability: String,
-}
-
-fn default_target_binding_availability() -> String {
-    "ready".to_string()
-}
-
-fn normalize_provider_selection_source(value: Option<String>) -> Option<String> {
-    value
-        .map(|source| source.trim().to_string())
-        .filter(|source| matches!(source.as_str(), "disk" | "managed"))
-}
-
-/// 当前选中的 Execution Target（Wave 4 / B.2 任务 2.3）。
-/// `provider_profile_id = None` 表示 default/local Provider 语义。
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct SharedSelectedReasoning {
-    pub(crate) effort: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct SharedSelectedTarget {
-    pub(crate) engine: EngineType,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) provider_profile_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) model_catalog_entry_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) model: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) reasoning: Option<SharedSelectedReasoning>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) provider_profile_name_snapshot: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) provider_profile_source: Option<String>,
-}
-
-fn normalize_shared_selected_target(mut target: SharedSelectedTarget) -> SharedSelectedTarget {
-    target.provider_profile_id = target
-        .provider_profile_id
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty());
-    target.model_catalog_entry_id = target
-        .model_catalog_entry_id
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty());
-    target.model = target
-        .model
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty());
-    target.reasoning = target.reasoning.and_then(|mut reasoning| {
-        reasoning.effort = reasoning.effort.trim().to_string();
-        (!reasoning.effort.is_empty()).then_some(reasoning)
-    });
-    target.provider_profile_name_snapshot = target
-        .provider_profile_name_snapshot
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty());
-    target.provider_profile_source =
-        normalize_provider_selection_source(target.provider_profile_source);
-    target
-}
-
-fn legacy_engine_only_selected_target(engine: EngineType) -> SharedSelectedTarget {
-    SharedSelectedTarget {
-        engine,
-        provider_profile_id: None,
-        model_catalog_entry_id: None,
-        model: None,
-        reasoning: None,
-        provider_profile_name_snapshot: None,
-        provider_profile_source: None,
-    }
-}
-
-fn is_legacy_engine_only_selected_target(target: &SharedSelectedTarget) -> bool {
-    target.provider_profile_id.is_none()
-        && target.model_catalog_entry_id.is_none()
-        && target.model.is_none()
-        && target.reasoning.is_none()
-        && target.provider_profile_name_snapshot.is_none()
-        && target.provider_profile_source.is_none()
-}
-
-fn validate_resolved_shared_selected_target(target: &SharedSelectedTarget) -> Result<(), String> {
-    let engine = ensure_supported_shared_session_engine(target.engine)?;
-    let provider_profile_id = target.provider_profile_id.as_deref();
-    let expected_source = if provider_profile_id.is_some() {
-        "managed"
-    } else {
-        "disk"
-    };
-    if target.provider_profile_source.as_deref() != Some(expected_source) {
-        return Err(format!(
-            "invalid-shared-target: provider source must be '{expected_source}'"
-        ));
-    }
-    if target.provider_profile_name_snapshot.is_none() {
-        return Err("invalid-shared-target: provider name snapshot is required".to_string());
-    }
-    if target.model_catalog_entry_id.is_none() || target.model.is_none() {
-        return Err(
-            "invalid-shared-target: modelCatalogEntryId and runtime model are required".to_string(),
-        );
-    }
-    let models = match provider_profile_id {
-        Some(provider_profile_id) => crate::engine::status::get_provider_scoped_engine_models(
-            engine,
-            Some(provider_profile_id),
-        )?,
-        None => crate::engine::status::get_local_engine_models_for_validation(engine),
-    };
-    // 与 shared_session_v2::validate_execution_target 同策：Qoder 模型目录是 ACP
-    // runtime-only（无静态 fallback roster），选择/持久化路径不得因目录不可得硬失败；
-    // catalog 可用时仍交叉校验 entry/model pair。
-    let models = match (engine, models) {
-        (EngineType::Qoder, None) => Vec::new(),
-        (_, Some(models)) => models,
-        (_, None) => {
-            return Err(format!(
-                "invalid-shared-target: model catalog is unavailable for {} provider {}",
-                engine.icon(),
-                provider_profile_id.unwrap_or("default")
-            ));
-        }
-    };
-    // 不限制用户模型名：catalog 未登记的自定义模型也允许保存为 next-send target。
-    crate::engine::status::validate_model_catalog_pair(
-        target.model_catalog_entry_id.as_deref(),
-        target.model.as_deref(),
-        &models,
-        crate::engine::status::UnlistedRuntimeModelPolicy::Allow,
-    )
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct SharedSessionMeta {
-    #[serde(default = "default_shared_session_schema_version")]
-    pub(crate) schema_version: u32,
-    pub(crate) id: String,
-    pub(crate) workspace_id: String,
-    pub(crate) title: String,
-    pub(crate) created_at: u64,
-    pub(crate) updated_at: u64,
-    /// V0 字段：保留写（= `selected_target.engine`），供回滚与旧版本读取。
-    selected_engine: EngineType,
-    /// Wave 4 起持久化；旧 meta 缺失时由 sanitize 从 `selected_engine` 迁移。
-    #[serde(default)]
-    pub(crate) selected_target: Option<SharedSelectedTarget>,
-    pub(crate) last_turn_seq: u64,
-    pub(crate) bindings_by_engine: HashMap<EngineType, SharedEngineBinding>,
-    /// Wave 4 起持久化；旧 meta 缺失时由 sanitize 从 `bindings_by_engine` 迁移。
-    #[serde(default)]
-    pub(crate) bindings_by_target: HashMap<String, SharedTargetBindingMeta>,
-}
-
-fn default_shared_session_schema_version() -> u32 {
-    SHARED_SESSION_SCHEMA_VERSION
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct SharedSessionSnapshotEntry {
-    kind: String,
-    created_at: u64,
-    selected_engine: EngineType,
-    last_turn_seq: u64,
-    pub(crate) items: Vec<Value>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct SharedSessionSummary {
-    pub(crate) id: String,
-    pub(crate) thread_id: String,
-    pub(crate) title: String,
-    pub(crate) created_at: u64,
-    pub(crate) updated_at: u64,
-    pub(crate) selected_engine: EngineType,
-    pub(crate) thread_kind: String,
-    pub(crate) engine_source: EngineType,
-    pub(crate) selected_engine_label: String,
-    pub(crate) native_thread_ids: Vec<String>,
-}
-
 #[derive(Debug, Clone, Serialize)]
 #[serde(rename_all = "camelCase")]
 struct SharedSessionLoadPayload {
@@ -288,537 +102,6 @@ struct SharedSessionLoadPayload {
     selected_target: Option<SharedSelectedTarget>,
     items: Vec<Value>,
     updated_at: u64,
-}
-
-struct SharedStoreFileLock {
-    path: PathBuf,
-}
-
-impl Drop for SharedStoreFileLock {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_file(&self.path);
-    }
-}
-
-fn shared_store_lock_file_path(path: &Path) -> PathBuf {
-    let extension = path
-        .extension()
-        .and_then(|value| value.to_str())
-        .map(|value| format!("{value}.lock"))
-        .unwrap_or_else(|| "lock".to_string());
-    path.with_extension(extension)
-}
-
-fn is_shared_store_lock_stale(lock_path: &Path) -> bool {
-    let metadata = match std::fs::metadata(lock_path) {
-        Ok(metadata) => metadata,
-        Err(_) => return false,
-    };
-    let modified_at = match metadata.modified() {
-        Ok(modified_at) => modified_at,
-        Err(_) => return false,
-    };
-    match modified_at.elapsed() {
-        Ok(elapsed) => elapsed > SHARED_STORE_LOCK_STALE_TIMEOUT,
-        Err(_) => false,
-    }
-}
-
-fn acquire_shared_store_lock(path: &Path) -> Result<SharedStoreFileLock, String> {
-    let lock_path = shared_store_lock_file_path(path);
-    if let Some(parent) = lock_path.parent() {
-        std::fs::create_dir_all(parent).map_err(|error| error.to_string())?;
-    }
-
-    let deadline = Instant::now() + SHARED_STORE_LOCK_WAIT_TIMEOUT;
-    loop {
-        match std::fs::OpenOptions::new()
-            .create_new(true)
-            .write(true)
-            .open(&lock_path)
-        {
-            Ok(mut file) => {
-                let _ = writeln!(file, "pid={}", std::process::id());
-                return Ok(SharedStoreFileLock { path: lock_path });
-            }
-            Err(error) if error.kind() == std::io::ErrorKind::AlreadyExists => {
-                if is_shared_store_lock_stale(&lock_path) {
-                    let _ = std::fs::remove_file(&lock_path);
-                    continue;
-                }
-                if Instant::now() >= deadline {
-                    return Err(format!(
-                        "Timed out waiting for shared session file lock: {}",
-                        lock_path.display()
-                    ));
-                }
-                thread::sleep(SHARED_STORE_LOCK_RETRY_INTERVAL);
-            }
-            Err(error) => return Err(error.to_string()),
-        }
-    }
-}
-
-fn with_shared_store_lock<T>(
-    path: &Path,
-    op: impl FnOnce() -> Result<T, String>,
-) -> Result<T, String> {
-    let _lock_guard = acquire_shared_store_lock(path)?;
-    op()
-}
-
-fn write_string_atomically(path: &Path, content: &str) -> Result<(), String> {
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent).map_err(|error| error.to_string())?;
-    }
-    let parent = path
-        .parent()
-        .ok_or_else(|| format!("Shared session path has no parent: {}", path.display()))?;
-    let filename = path
-        .file_name()
-        .and_then(|value| value.to_str())
-        .ok_or_else(|| {
-            format!(
-                "Shared session path has invalid filename: {}",
-                path.display()
-            )
-        })?;
-    let temp_path = parent.join(format!(".{filename}.{}.tmp", Uuid::new_v4()));
-    let mut temp_file = std::fs::OpenOptions::new()
-        .create_new(true)
-        .write(true)
-        .open(&temp_path)
-        .map_err(|error| error.to_string())?;
-    temp_file
-        .write_all(content.as_bytes())
-        .map_err(|error| error.to_string())?;
-    temp_file.sync_all().map_err(|error| error.to_string())?;
-
-    #[cfg(target_os = "windows")]
-    if path.exists() {
-        std::fs::remove_file(path).map_err(|error| error.to_string())?;
-    }
-
-    if let Err(error) = std::fs::rename(&temp_path, path) {
-        let _ = std::fs::remove_file(&temp_path);
-        return Err(error.to_string());
-    }
-    Ok(())
-}
-
-pub(crate) fn now_millis() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_else(|_| Duration::from_millis(0))
-        .as_millis() as u64
-}
-
-fn shared_sessions_root_dir() -> Result<PathBuf, String> {
-    Ok(app_paths::app_home_dir()?.join(SHARED_SESSIONS_DIRNAME))
-}
-
-fn workspace_shared_sessions_dir(workspace_id: &str) -> Result<PathBuf, String> {
-    Ok(shared_sessions_root_dir()?.join(workspace_id))
-}
-
-fn shared_session_dir(workspace_id: &str, shared_session_id: &str) -> Result<PathBuf, String> {
-    Ok(workspace_shared_sessions_dir(workspace_id)?.join(shared_session_id))
-}
-
-fn shared_session_meta_path(
-    workspace_id: &str,
-    shared_session_id: &str,
-) -> Result<PathBuf, String> {
-    Ok(shared_session_dir(workspace_id, shared_session_id)?.join("meta.json"))
-}
-
-fn shared_session_log_path(workspace_id: &str, shared_session_id: &str) -> Result<PathBuf, String> {
-    Ok(shared_session_dir(workspace_id, shared_session_id)?.join("log.jsonl"))
-}
-
-pub(crate) fn shared_session_projection_source(
-    workspace_id: &str,
-    thread_id: &str,
-) -> Result<(String, PathBuf), String> {
-    let shared_session_id = parse_shared_session_id(thread_id)?;
-    let log_path = shared_session_log_path(workspace_id, &shared_session_id)?;
-    Ok((shared_session_id, log_path))
-}
-
-fn shared_thread_id(shared_session_id: &str) -> String {
-    format!("shared:{shared_session_id}")
-}
-
-fn is_safe_shared_session_storage_id(value: &str) -> bool {
-    !value.is_empty()
-        && value != "."
-        && value != ".."
-        && value
-            .chars()
-            .all(|ch| ch.is_ascii_alphanumeric() || ch == '-' || ch == '_')
-}
-
-pub(crate) fn parse_shared_session_id(thread_id: &str) -> Result<String, String> {
-    let normalized = thread_id.trim();
-    if let Some(rest) = normalized.strip_prefix("shared:") {
-        let shared_session_id = rest.trim();
-        if is_safe_shared_session_storage_id(shared_session_id) {
-            return Ok(shared_session_id.to_string());
-        }
-    }
-    Err(format!("Invalid shared session thread id: {thread_id}"))
-}
-
-fn validate_shared_native_thread_id(value: &str) -> Result<String, String> {
-    let normalized = value.trim();
-    if normalized.is_empty() {
-        Err("Shared session native thread id cannot be empty".to_string())
-    } else {
-        Ok(normalized.to_string())
-    }
-}
-
-fn canonical_shared_native_thread_id(
-    engine: EngineType,
-    provider_profile_id: Option<&str>,
-    native_thread_id: &str,
-) -> String {
-    let native_thread_id = native_thread_id.trim();
-    if native_thread_id.is_empty()
-        || engine != EngineType::Qoder
-        || is_pending_shared_binding_thread_id(engine, native_thread_id)
-    {
-        return native_thread_id.to_string();
-    }
-    match crate::engine::qoder_provider_profile::canonical_qoder_native_session_id(
-        native_thread_id,
-        provider_profile_id,
-    ) {
-        Ok(identity) => identity,
-        Err(error) => {
-            // 保留无法解释的旧值，避免 metadata sanitize 造成数据丢失；下游
-            // visibility 会拒绝把它展开为跨 distribution 的 bare raw alias。
-            log::warn!(
-                "[shared_sessions] retained malformed Qoder native binding `{native_thread_id}`: {error}"
-            );
-            native_thread_id.to_string()
-        }
-    }
-}
-
-pub(crate) fn is_pending_shared_binding_thread_id(engine: EngineType, thread_id: &str) -> bool {
-    let normalized = thread_id.trim();
-    if normalized.is_empty() {
-        return true;
-    }
-    match engine {
-        EngineType::Claude => normalized.starts_with("claude-pending-shared-"),
-        EngineType::Codex => normalized.starts_with("codex-pending-shared-"),
-        EngineType::Kimi => normalized.starts_with("kimi-pending-shared-"),
-        EngineType::Pi => normalized.starts_with("pi-pending-shared-"),
-        EngineType::Grok => normalized.starts_with("grok-pending-shared-"),
-        EngineType::OpenCode => normalized.starts_with("opencode-pending-shared-"),
-        EngineType::Qoder => normalized.starts_with("qoder-pending-shared-"),
-        EngineType::Gemini | EngineType::Dsh => false,
-    }
-}
-
-pub(crate) fn binding_uses_established_native_thread(engine: EngineType, thread_id: &str) -> bool {
-    let normalized = thread_id.trim();
-    if normalized.is_empty() || is_pending_shared_binding_thread_id(engine, normalized) {
-        return false;
-    }
-    // 兼容 `engine:{raw}` 与历史 raw id；strip 前缀后再判 pending。
-    let raw = match engine {
-        EngineType::Claude
-        | EngineType::Kimi
-        | EngineType::Pi
-        | EngineType::Grok
-        | EngineType::OpenCode
-        | EngineType::Dsh
-        | EngineType::Qoder => {
-            let prefix = format!("{}:", engine.icon());
-            normalized
-                .strip_prefix(prefix.as_str())
-                .unwrap_or(normalized)
-                .trim()
-        }
-        EngineType::Codex | EngineType::Gemini => normalized,
-    };
-    if raw.is_empty() || is_pending_shared_binding_thread_id(engine, raw) {
-        return false;
-    }
-    match engine {
-        EngineType::Claude => normalized.contains(':'),
-        EngineType::Codex
-        | EngineType::Kimi
-        | EngineType::Pi
-        | EngineType::Grok
-        | EngineType::OpenCode
-        | EngineType::Qoder => true,
-        EngineType::Gemini | EngineType::Dsh => false,
-    }
-}
-
-pub(crate) fn engine_binding_thread_id(engine: EngineType, seed: &str) -> String {
-    match engine {
-        EngineType::Claude => format!("claude-pending-shared-{seed}"),
-        EngineType::Codex => format!("codex-pending-shared-{seed}"),
-        EngineType::Kimi => format!("kimi-pending-shared-{seed}"),
-        EngineType::Pi => format!("pi-pending-shared-{seed}"),
-        EngineType::Grok => format!("grok-pending-shared-{seed}"),
-        EngineType::OpenCode => format!("opencode-pending-shared-{seed}"),
-        EngineType::Gemini => format!("gemini-pending-shared-{seed}"),
-        EngineType::Dsh => format!("dsh-pending-shared-{seed}"),
-        // Qoder Shared bindings retain their distribution identity; this id is only provisional
-        // until the corresponding native session is established.
-        EngineType::Qoder => format!("qoder-pending-shared-{seed}"),
-    }
-}
-
-/// Binding Key = Engine + ProviderProfile（Model 不进 Key）。
-/// 与前端 `bindingKeyOf` 保持一致：`{engine}:{provider|"default"}`。
-pub(crate) fn shared_target_binding_key(
-    engine: EngineType,
-    provider_profile_id: Option<&str>,
-) -> String {
-    let provider = provider_profile_id
-        .map(str::trim)
-        .filter(|value| !value.is_empty());
-    format!("{}:{}", engine.icon(), provider.unwrap_or("default"))
-}
-
-fn sanitize_shared_session_meta(meta: &mut SharedSessionMeta) {
-    meta.schema_version = SHARED_SESSION_SCHEMA_VERSION;
-    // 任务 2.3：`selectedEngine → selectedTarget` 迁移；selectedTarget 为权威，
-    // selected_engine 回落为 target.engine（V0 回滚读取兼容）。
-    let mut selected_target = match meta.selected_target.take() {
-        Some(mut target) => {
-            if !is_supported_shared_session_engine(target.engine) {
-                target.engine = normalize_shared_session_engine(target.engine);
-                target.provider_profile_id = None;
-                target.model_catalog_entry_id = None;
-                target.model = None;
-                target.reasoning = None;
-                target.provider_profile_name_snapshot = None;
-                target.provider_profile_source = None;
-            }
-            target.provider_profile_id = target
-                .provider_profile_id
-                .map(|value| value.trim().to_string())
-                .filter(|value| !value.is_empty());
-            target
-        }
-        None => SharedSelectedTarget {
-            engine: normalize_shared_session_engine(meta.selected_engine),
-            provider_profile_id: None,
-            model_catalog_entry_id: None,
-            model: None,
-            reasoning: None,
-            provider_profile_name_snapshot: None,
-            provider_profile_source: None,
-        },
-    };
-    selected_target.model_catalog_entry_id = selected_target
-        .model_catalog_entry_id
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty());
-    selected_target.model = selected_target
-        .model
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty());
-    selected_target.reasoning = selected_target.reasoning.and_then(|mut reasoning| {
-        reasoning.effort = reasoning.effort.trim().to_string();
-        (!reasoning.effort.is_empty()).then_some(reasoning)
-    });
-    selected_target.provider_profile_name_snapshot = selected_target
-        .provider_profile_name_snapshot
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty());
-    selected_target.provider_profile_source =
-        normalize_provider_selection_source(selected_target.provider_profile_source);
-    meta.selected_engine = selected_target.engine;
-    meta.selected_target = Some(selected_target);
-    meta.bindings_by_engine
-        .retain(|engine, _| is_supported_shared_session_engine(*engine));
-    for (engine, binding) in meta.bindings_by_engine.iter_mut() {
-        binding.engine = *engine;
-        binding.native_thread_id =
-            canonical_shared_native_thread_id(*engine, None, &binding.native_thread_id);
-    } // B.2 迁移：旧 `bindings_by_engine` 归位到 default-provider 语义。
-      // V0 仍是 default binding 身份字段的权威来源（回滚兼容），
-      // 因此 default key 的身份字段以 engine binding 为准做覆盖式同步；
-      // managed-provider 条目（provider_profile_id != None）不受此影响。
-    meta.bindings_by_target.retain(|key, binding| {
-        key == &binding.binding_key && is_supported_shared_session_engine(binding.engine)
-    });
-    for binding in meta.bindings_by_target.values_mut() {
-        binding.native_thread_id = canonical_shared_native_thread_id(
-            binding.engine,
-            binding.provider_profile_id.as_deref(),
-            &binding.native_thread_id,
-        );
-    }
-    for (engine, binding) in meta.bindings_by_engine.iter() {
-        let key = shared_target_binding_key(*engine, None);
-        match meta.bindings_by_target.get_mut(&key) {
-            Some(target) => {
-                target.engine = *engine;
-                target.provider_profile_id = None;
-                target.native_thread_id = binding.native_thread_id.clone();
-                target.created_at = binding.created_at;
-                target.last_used_at = binding.last_used_at;
-                target.last_synced_turn_seq = binding.last_synced_turn_seq;
-            }
-            None => {
-                meta.bindings_by_target.insert(
-                    key.clone(),
-                    SharedTargetBindingMeta {
-                        binding_key: key,
-                        engine: *engine,
-                        provider_profile_id: None,
-                        native_thread_id: binding.native_thread_id.clone(),
-                        created_at: binding.created_at,
-                        last_used_at: binding.last_used_at,
-                        last_synced_turn_seq: binding.last_synced_turn_seq,
-                        availability: default_target_binding_availability(),
-                    },
-                );
-            }
-        }
-    }
-    // default binding 在 engine map 中已不存在时，target map 也不应残留（例如 sanitize 剔除不支持 engine）。
-    let live_default_keys: std::collections::HashSet<String> = meta
-        .bindings_by_engine
-        .keys()
-        .map(|engine| shared_target_binding_key(*engine, None))
-        .collect();
-    meta.bindings_by_target.retain(|key, binding| {
-        binding.provider_profile_id.is_some() || live_default_keys.contains(key)
-    });
-}
-
-/// 更新选中 Target（同时写 V0 `selected_engine` 保持回滚兼容）。
-pub(crate) fn select_meta_target(
-    meta: &mut SharedSessionMeta,
-    engine: EngineType,
-    provider_profile_id: Option<String>,
-) {
-    let provider_profile_id = provider_profile_id
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty());
-    let preserved = meta.selected_target.take().filter(|target| {
-        target.engine == engine && target.provider_profile_id == provider_profile_id
-    });
-    meta.selected_engine = engine;
-    meta.selected_target = Some(preserved.unwrap_or(SharedSelectedTarget {
-        engine,
-        provider_profile_id,
-        model_catalog_entry_id: None,
-        model: None,
-        reasoning: None,
-        provider_profile_name_snapshot: None,
-        provider_profile_source: None,
-    }));
-}
-
-fn apply_selected_target_selection(
-    root: &mut Value,
-    target: &SharedSelectedTarget,
-    updated_at: u64,
-) -> Result<(), String> {
-    let object = root
-        .as_object_mut()
-        .ok_or_else(|| "Shared session metadata must be a JSON object".to_string())?;
-    object.insert(
-        "selectedEngine".to_string(),
-        serde_json::to_value(target.engine).map_err(|error| error.to_string())?,
-    );
-    object.insert(
-        "selectedTarget".to_string(),
-        serde_json::to_value(target).map_err(|error| error.to_string())?,
-    );
-    object.insert("updatedAt".to_string(), json!(updated_at));
-    Ok(())
-}
-
-fn write_shared_session_selection(
-    workspace_id: &str,
-    shared_session_id: &str,
-    target: &SharedSelectedTarget,
-    updated_at: u64,
-    writer: &SharedEventWriter,
-) -> Result<SharedSelectedTarget, String> {
-    let path = shared_session_meta_path(workspace_id, shared_session_id)?;
-    with_shared_store_lock(&path, || {
-        let raw = std::fs::read_to_string(&path).map_err(|error| error.to_string())?;
-        let mut root: Value = serde_json::from_str(&raw).map_err(|error| error.to_string())?;
-        let selected_target = if is_legacy_engine_only_selected_target(target) {
-            let mut meta: SharedSessionMeta =
-                serde_json::from_value(root.clone()).map_err(|error| error.to_string())?;
-            sanitize_shared_session_meta(&mut meta);
-            resolve_shared_selection_update(&mut meta, target)
-        } else {
-            target.clone()
-        };
-        apply_selected_target_selection(&mut root, &selected_target, updated_at)?;
-        let updated_raw = serde_json::to_string_pretty(&root).map_err(|error| error.to_string())?;
-        write_string_atomically(&path, &updated_raw)?;
-        if let Err(error) =
-            upsert_v2_selected_target(writer, shared_session_id, &selected_target, updated_at)
-        {
-            let rollback = write_string_atomically(&path, &raw);
-            return Err(match rollback {
-                Ok(()) => error,
-                Err(rollback_error) => {
-                    format!("{error}; legacy metadata rollback also failed: {rollback_error}")
-                }
-            });
-        }
-        Ok(selected_target)
-    })
-}
-
-fn upsert_v2_selected_target(
-    writer: &SharedEventWriter,
-    shared_session_id: &str,
-    target: &SharedSelectedTarget,
-    updated_at: u64,
-) -> Result<(), String> {
-    let selected_target_json = serde_json::to_string(target).map_err(|error| error.to_string())?;
-    writer
-        .upsert_session_target(&SessionTargetUpdate {
-            session_id: shared_session_id.to_string(),
-            schema_version: SHARED_SESSION_SCHEMA_VERSION,
-            selected_target_json,
-            updated_at: updated_at as i64,
-        })
-        .map_err(|error| error.to_string())
-}
-
-fn select_meta_engine_compat(meta: &mut SharedSessionMeta, engine: EngineType) {
-    if meta
-        .selected_target
-        .as_ref()
-        .is_some_and(|target| target.engine == engine)
-    {
-        meta.selected_engine = engine;
-        return;
-    }
-    select_meta_target(meta, engine, None);
-}
-
-fn resolve_shared_selection_update(
-    meta: &mut SharedSessionMeta,
-    requested_target: &SharedSelectedTarget,
-) -> SharedSelectedTarget {
-    if !is_legacy_engine_only_selected_target(requested_target) {
-        return requested_target.clone();
-    }
-    select_meta_engine_compat(meta, requested_target.engine);
-    meta.selected_target
-        .clone()
-        .unwrap_or_else(|| legacy_engine_only_selected_target(requested_target.engine))
 }
 
 async fn ensure_shared_session_native_binding(
@@ -981,396 +264,6 @@ fn touch_shared_binding(
             binding.last_synced_turn_seq = synced;
         }
     }
-}
-
-pub(crate) fn read_shared_session_meta(
-    workspace_id: &str,
-    shared_session_id: &str,
-) -> Result<SharedSessionMeta, String> {
-    let path = shared_session_meta_path(workspace_id, shared_session_id)?;
-    let raw = std::fs::read_to_string(&path).map_err(|error| error.to_string())?;
-    let mut meta: SharedSessionMeta =
-        serde_json::from_str(&raw).map_err(|error| error.to_string())?;
-    sanitize_shared_session_meta(&mut meta);
-    Ok(meta)
-}
-
-/// Filesystem-only Shared ownership seed (V0 metadata). Does not touch EventWriter.
-#[derive(Debug, Clone, Default)]
-pub(crate) struct WorkspaceSharedOwnershipSeed {
-    pub session_ids: Vec<String>,
-    pub native_ids: Vec<String>,
-    pub skipped_meta: usize,
-}
-
-pub(crate) fn load_workspace_shared_ownership_seed(
-    workspace_id: &str,
-) -> Result<WorkspaceSharedOwnershipSeed, String> {
-    let directory = workspace_shared_sessions_dir(workspace_id)?;
-    if !directory.exists() {
-        return Ok(WorkspaceSharedOwnershipSeed::default());
-    }
-    load_seed_from_shared_sessions_dir(&directory)
-}
-
-/// Iterate one `shared-sessions/<workspace_id>` directory. Split out for
-/// tests (temp dir) — behavior must stay identical to reading via
-/// workspace_id paths.
-fn load_seed_from_shared_sessions_dir(
-    directory: &std::path::Path,
-) -> Result<WorkspaceSharedOwnershipSeed, String> {
-    let mut seed = WorkspaceSharedOwnershipSeed::default();
-    for entry in std::fs::read_dir(directory).map_err(|error| error.to_string())? {
-        let entry = entry.map_err(|error| error.to_string())?;
-        let file_type = entry.file_type().map_err(|error| error.to_string())?;
-        if !file_type.is_dir() {
-            continue;
-        }
-        let shared_session_id = entry.file_name().to_string_lossy().to_string();
-        if shared_session_id.trim().is_empty() {
-            continue;
-        }
-        let meta_path = entry.path().join("meta.json");
-        // meta.json 整个缺失（崩溃/中断创建的残留空目录）不是真实 session，
-        // 静默跳过；只有 meta.json 存在但读不出/解析失败才按保守策略记
-        // skipped_meta（→ projection unavailable，前端 defer 全部 native
-        // 行直到 hide 验证）。此前残留空目录会让 skipped_meta>0 →
-        // visibility 恒 unavailable → 侧栏 first-paint early paint 全部
-        // 被 defer，首段必空（2026-08-26 实证：7 个 workspace 42 个残留
-        // 空目录，reason=legacy-meta-skipped:N）。
-        if !meta_path.exists() {
-            continue;
-        }
-        let meta = std::fs::read_to_string(&meta_path)
-            .ok()
-            .and_then(|raw| serde_json::from_str::<SharedSessionMeta>(&raw).ok());
-        let Some(mut meta) = meta else {
-            seed.skipped_meta += 1;
-            continue;
-        };
-        sanitize_shared_session_meta(&mut meta);
-        seed.session_ids.push(meta.id.clone());
-        for (engine, binding) in &meta.bindings_by_engine {
-            let native_id =
-                canonical_shared_native_thread_id(*engine, None, &binding.native_thread_id);
-            let native_id = native_id.trim();
-            if !native_id.is_empty() {
-                seed.native_ids.push(native_id.to_string());
-            }
-        }
-        for binding in meta.bindings_by_target.values() {
-            let native_id = canonical_shared_native_thread_id(
-                binding.engine,
-                binding.provider_profile_id.as_deref(),
-                &binding.native_thread_id,
-            );
-            let native_id = native_id.trim();
-            if !native_id.is_empty() {
-                seed.native_ids.push(native_id.to_string());
-            }
-        }
-    }
-    seed.session_ids.sort();
-    seed.session_ids.dedup();
-    seed.native_ids.sort();
-    seed.native_ids.dedup();
-    Ok(seed)
-}
-
-pub(crate) fn write_shared_session_meta(meta: &SharedSessionMeta) -> Result<(), String> {
-    let path = shared_session_meta_path(&meta.workspace_id, &meta.id)?;
-    with_shared_store_lock(&path, || {
-        let mut sanitized = meta.clone();
-        sanitize_shared_session_meta(&mut sanitized);
-        let raw = serde_json::to_string_pretty(&sanitized).map_err(|error| error.to_string())?;
-        write_string_atomically(&path, &raw)
-    })
-}
-
-fn append_shared_session_log_entry(
-    workspace_id: &str,
-    shared_session_id: &str,
-    entry: &SharedSessionSnapshotEntry,
-) -> Result<(), String> {
-    let path = shared_session_log_path(workspace_id, shared_session_id)?;
-    with_shared_store_lock(&path, || {
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent).map_err(|error| error.to_string())?;
-        }
-        let serialized = serde_json::to_string(entry).map_err(|error| error.to_string())?;
-        let mut file = std::fs::OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(&path)
-            .map_err(|error| error.to_string())?;
-        writeln!(file, "{serialized}").map_err(|error| error.to_string())?;
-        file.sync_all().map_err(|error| error.to_string())?;
-        Ok(())
-    })
-}
-
-pub(crate) fn read_latest_shared_session_snapshot(
-    workspace_id: &str,
-    shared_session_id: &str,
-) -> Result<Option<SharedSessionSnapshotEntry>, String> {
-    let path = shared_session_log_path(workspace_id, shared_session_id)?;
-    if !path.exists() {
-        return Ok(None);
-    }
-    let content = std::fs::read_to_string(&path).map_err(|error| error.to_string())?;
-    let latest = content
-        .lines()
-        .filter(|line| !line.trim().is_empty())
-        .filter_map(|line| serde_json::from_str::<SharedSessionSnapshotEntry>(line).ok())
-        .last();
-    Ok(latest)
-}
-
-pub(crate) fn list_workspace_shared_sessions(
-    workspace_id: &str,
-    event_writer: Option<&crate::shared_event_log::SharedEventWriter>,
-    event_log_path: Option<&Path>,
-) -> Result<Vec<SharedSessionSummary>, String> {
-    let directory = workspace_shared_sessions_dir(workspace_id)?;
-    if !directory.exists() {
-        return Ok(Vec::new());
-    }
-    let mut session_metas = Vec::new();
-    for entry in std::fs::read_dir(&directory).map_err(|error| error.to_string())? {
-        let entry = entry.map_err(|error| error.to_string())?;
-        let file_type = entry.file_type().map_err(|error| error.to_string())?;
-        if !file_type.is_dir() {
-            continue;
-        }
-        let shared_session_id = entry.file_name().to_string_lossy().to_string();
-        let meta = match read_shared_session_meta(workspace_id, &shared_session_id) {
-            Ok(meta) => meta,
-            Err(_) => continue,
-        };
-        session_metas.push(meta);
-    }
-
-    let v2_native_thread_ids_by_session = if event_writer.is_none() {
-        let shared_session_ids = session_metas
-            .iter()
-            .map(|meta| meta.id.clone())
-            .collect::<Vec<_>>();
-        match event_log_path {
-            Some(path) => match collect_v2_shared_binding_ids_by_session(path, &shared_session_ids)
-            {
-                Ok(ids_by_session) => ids_by_session,
-                Err(error) => {
-                    log::warn!(
-                        "[shared_sessions.list_workspace_shared_sessions] read-only V2 binding recovery failed for workspace {}: {}",
-                        workspace_id,
-                        error
-                    );
-                    BTreeMap::new()
-                }
-            },
-            None => BTreeMap::new(),
-        }
-    } else {
-        BTreeMap::new()
-    };
-
-    let mut summaries = Vec::with_capacity(session_metas.len());
-    for meta in session_metas {
-        let mut native_thread_ids = meta
-            .bindings_by_engine
-            .iter()
-            .map(|(engine, binding)| {
-                canonical_shared_native_thread_id(*engine, None, &binding.native_thread_id)
-            })
-            .collect::<Vec<_>>();
-        native_thread_ids.extend(meta.bindings_by_target.values().map(|binding| {
-            canonical_shared_native_thread_id(
-                binding.engine,
-                binding.provider_profile_id.as_deref(),
-                &binding.native_thread_id,
-            )
-        }));
-        if let Some(writer) = event_writer {
-            native_thread_ids.extend(
-                writer
-                    .binding_states_for_session(&meta.id)
-                    .map_err(|error| error.to_string())?
-                    .into_iter()
-                    .filter_map(|binding| {
-                        binding.native_session_id.map(|native_session_id| {
-                            if binding.engine == EngineType::Qoder.icon() {
-                                canonical_shared_native_thread_id(
-                                    EngineType::Qoder,
-                                    binding.provider_profile_id.as_deref(),
-                                    &native_session_id,
-                                )
-                            } else {
-                                native_session_id
-                            }
-                        })
-                    })
-                    .filter(|native_session_id| !native_session_id.trim().is_empty()),
-            );
-        } else if let Some(v2_native_thread_ids) = v2_native_thread_ids_by_session.get(&meta.id) {
-            native_thread_ids.extend(v2_native_thread_ids.iter().cloned());
-        }
-        native_thread_ids.sort();
-        native_thread_ids.dedup();
-        summaries.push(SharedSessionSummary {
-            id: meta.id.clone(),
-            thread_id: shared_thread_id(&meta.id),
-            title: meta.title.clone(),
-            created_at: meta.created_at,
-            updated_at: meta.updated_at,
-            selected_engine: meta.selected_engine,
-            thread_kind: "shared".to_string(),
-            engine_source: meta.selected_engine,
-            selected_engine_label: meta.selected_engine.display_name().to_string(),
-            native_thread_ids,
-        });
-    }
-    summaries.sort_by(|left, right| right.updated_at.cmp(&left.updated_at));
-    Ok(summaries)
-}
-
-fn extract_first_user_title(items: &[Value]) -> Option<String> {
-    for item in items {
-        let role = item
-            .get("role")
-            .and_then(Value::as_str)
-            .unwrap_or_default()
-            .trim()
-            .to_ascii_lowercase();
-        if role != "user" {
-            continue;
-        }
-        let text = item
-            .get("text")
-            .and_then(Value::as_str)
-            .unwrap_or_default()
-            .trim();
-        if text.is_empty() {
-            continue;
-        }
-        let normalized = text.lines().next().unwrap_or(text).trim();
-        if normalized.is_empty() {
-            continue;
-        }
-        let title = if normalized.chars().count() > 32 {
-            format!("{}...", normalized.chars().take(32).collect::<String>())
-        } else {
-            normalized.to_string()
-        };
-        return Some(title);
-    }
-    None
-}
-
-fn count_user_turns(items: &[Value]) -> u64 {
-    items
-        .iter()
-        .filter(|item| {
-            item.get("kind").and_then(Value::as_str) == Some("message")
-                && item.get("role").and_then(Value::as_str) == Some("user")
-        })
-        .count() as u64
-}
-
-fn build_delta_sync_projection(items: &[Value], from_turn_seq: u64) -> Option<(String, bool)> {
-    if items.is_empty() {
-        return None;
-    }
-    let mut turn_index = 0_u64;
-    let mut current_user: Option<String> = None;
-    let mut collected: Vec<String> = Vec::new();
-
-    for item in items {
-        let kind = item.get("kind").and_then(Value::as_str).unwrap_or_default();
-        if kind != "message" {
-            continue;
-        }
-        let role = item.get("role").and_then(Value::as_str).unwrap_or_default();
-        let text = item
-            .get("text")
-            .and_then(Value::as_str)
-            .unwrap_or_default()
-            .trim()
-            .replace('\n', " ");
-        if role == "user" {
-            turn_index += 1;
-            current_user = if text.is_empty() { None } else { Some(text) };
-            continue;
-        }
-        if role == "assistant" && turn_index > from_turn_seq {
-            let engine = item
-                .get("engineSource")
-                .and_then(Value::as_str)
-                .unwrap_or("assistant")
-                .trim()
-                .to_string();
-            if !text.is_empty() {
-                if let Some(user_text) = current_user.take() {
-                    collected.push(format!(
-                        "Turn {turn_index}\nUser: {user_text}\n{engine}: {text}"
-                    ));
-                }
-            }
-        }
-    }
-
-    if collected.is_empty() {
-        return None;
-    }
-
-    let mut merged = String::from(
-        "Shared session context sync. Continue from these recent turns before answering the new request:\n\n",
-    );
-    let retained_from = collected.len().saturating_sub(MAX_DELTA_SYNC_TURNS);
-    let mut truncated = false;
-    for block in &collected[retained_from..] {
-        let remaining = MAX_DELTA_SYNC_CHARS.saturating_sub(merged.chars().count() + 2);
-        if remaining == 0 {
-            truncated = true;
-            break;
-        }
-        let block_chars = block.chars().count();
-        merged.extend(block.chars().take(remaining));
-        merged.push_str("\n\n");
-        if block_chars > remaining {
-            truncated = true;
-            break;
-        }
-    }
-    Some((merged.trim_end().to_string(), truncated))
-}
-
-fn build_delta_sync_prefix(items: &[Value], from_turn_seq: u64) -> Option<String> {
-    build_delta_sync_projection(items, from_turn_seq).map(|(projection, _)| projection)
-}
-
-pub(crate) fn inspect_shared_context_projection(
-    items: &[Value],
-    from_turn_seq: u64,
-) -> Vec<String> {
-    let pending_turns = count_user_turns(items).saturating_sub(from_turn_seq);
-    let mut omissions = Vec::new();
-    if pending_turns as usize > MAX_DELTA_SYNC_TURNS {
-        omissions.push(format!(
-            "{} older turn(s) omitted by the {}-turn context limit",
-            pending_turns as usize - MAX_DELTA_SYNC_TURNS,
-            MAX_DELTA_SYNC_TURNS
-        ));
-    }
-    let projection_truncated = build_delta_sync_projection(items, from_turn_seq)
-        .map(|(_, truncated)| truncated)
-        .unwrap_or(false);
-    if projection_truncated {
-        omissions.push(format!(
-            "context truncated at the {}-character limit",
-            MAX_DELTA_SYNC_CHARS
-        ));
-    }
-    omissions
 }
 
 pub(crate) fn shared_binding_synced_sequence(
